+++ src/elements/segment.jsx
@@ -3,13 +3,14 @@ module.exports = function (React) {
 
   var ClassGenerator = require('../mixins/classGenerator.js')(React);
   var ColorSelector = require('../mixins/colorSelector.js')(React);
+  var StateSelector  = require('../mixins/stateSelector.js')(React);
   var Unit = require('../commons/unit.jsx')(React);
 
   var defaultClassName = 'ui segment';
 
   var Segment = React.createClass({
 
-    mixins: [ClassGenerator, ColorSelector],
+    mixins: [ClassGenerator, ColorSelector, StateSelector],
 
     render: function () {
 
@@ -19,7 +20,9 @@ module.exports = function (React) {
         <Unit {...other}
           className={this.getClassName(defaultClassName)}
           type="div"
-          color={this.getColor()}>
+          color={this.getColor()}
+          disabled={this.getDisabled()}
+          loading={this.getLoading()}>
         </Unit>
       );
     }