+++ src/elements/step.jsx
@@ -2,21 +2,29 @@
 module.exports = function (React) {
 
   var ClassGenerator = require('../mixins/classGenerator.js')(React);
+  var StateSelector  = require('../mixins/stateSelector.js')(React);
+  var Unit = require('../commons/unit.jsx')(React);
 
   var defaultClassName = 'step';
 
   var Step = React.createClass({
 
-    mixins: [ClassGenerator],
+    mixins: [ClassGenerator, StateSelector],
 
     render: function () {
 
       var {className, ...other} = this.props;
 
       return (
-        <div {...other} className={this.getClassName(defaultClassName)} >
+        <Unit {...other}
+          className={this.getClassName(defaultClassName)}
+          type="div"
+          color="null"
+          active={this.getActive()}
+          completed={this.getCompleted()}
+          disabled={this.getDisabled()}>
           {this.props.children}
-        </div>
+        </Unit>
       );
     }
   });